TYPO3CR: * implemented Node.getReferences() and Node.getWeakReferences(), thanks to Matthias Hörmann
Implementation is based on a more general method in the storage backend to get all properties of a given type and value and optionally a name
Both that more general method and Node.get(Weak)References() are tested but the former is only tested with reference types and won't work
with types where the type-specific table uses more than just a value column to store the actual value (e.g. not for path or name values).