[+BUGFIX] TYPO3CR: Persist live Workspace if user workspace is created first
If the first request after initial setup of the demo site is not addressing
the live workspace but a user workspace, that user workspace is created
but doesn't have the live workspace as a base workspace.

This patch fixes this behavior and makes sure that the live workspace is
also created and persisted as expected.

Change-Id: I2e56e85026bbe292bc686eec9c3ce816e18fbe74