[~TASK] TYPO3CR: Implement tests for Workspace
This adds a test suite for the Workspace domain model assuring
the essential functionality. Code coverage is 100% for the
Workspace class.

While writing the tests, I also did some small refactoring of
the involved classes. Worth mentioning is the createNode() method
which now accepts the content type of the new node as the second
argument.

Relates to: #10622

Change-Id: I553582b9516368fd71085f70eac7635421c81979