[+BUGFIX] TYPO3CR: Make findByParentAndContentType() added objects aware
findByParentAndContentType() and countByParentAndContentType() were
not considering new nodes which were added during the current request
(and thus not yet persisted). This resulted in a wrong result which
in turned resulted in wrong index values for newly created nodes.

The most visible symptom was that the demo site was imported but only
few of the pages were visible in the site's menu because they had
conflicting index values.

Change-Id: I88bfcbede1e19831b5b5a141a7b6896381536093