[TASK] Ajdust code for Doctrine persistence
Adds some annotations and adjusts some code.

NodeRepository now handles added and removed objects internally, because
we rely on access to it for some workspace magic that needs to work
around some of our own design decisions (not yet persisted items are not
visible in queries).

Another big change is the introduction of a ContentObjectProxy to work
around the fact that Doctrine 2 cannot persist associations to
arbitrary objects but needs a target type. This is still not a real
solutions but more a hack...

Change-Id: Ia0c04ccb940a7894b72703f388f4c9e490330798