[BUGFIX] (Domain): NodeRepository can now order nodes with same sort index
Sometimes, multiple Node objects on the same level have the same
index (used for determining the order of the nodes). It is important
that a well-defined way to resolve these conflicts exist.

The current solution defines two sorting keys: First, it uses
the index, and secondly, it uses the Identifier to have a
deterministic ordering.

This patch is a prerequisite for making the HTML Content Element work.

Change-Id: I46fe9f32bbd1d4551651f46c98c7a9e1f4114404