[~TASK] Allow to switch off automatic workspace creation in getWorkspace()
Until now Context::getWorkspace() would always implicitly create a new
workspace if no workspace with the given name exists. This is not what
you want for example while checking if a node exists in the Node route
part handlers. A flag now allows to switch off this automatic workspace
creation.

Change-Id: I6334ac4457a6fc3f9af1c067bae81821100ac242