[TASK] isVisible() does not consider accessible status anymore
This partly reverts my earlier commit which let isVisible() check
if a node is accessible.

Now isVisible() will only tell something about the visibility status,
regardless of if the node _may_ be viewed or not. This must be tested
separately by calling isAccessible().

Additionally this change set includes functions for setting a flag
"inaccessibleContentShown" in the Context. This is for future use.

Change-Id: Ic81348427840f2cb66ae5bdd4f559d93141de094
Relates: #29538